backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gunicorn==21.2.0
python-dotenv==1.0.0
boto3==1.34.14
transformers==4.36.2
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.1.2+cpu
pydantic==2.5.3
python-multipart==0.0.6